requirements.txt
google-api-python-client==1.9.3
google-auth~=1.18.0
google-colab==1.0.0
google==2.0.3
imageio==2.8.0
imutils==0.5.3
ipykernel~=5.3.0
jupyter==1.0.0
matplotlib==3.2.2
notebook~=6.0.3
numpy==1.19.0
Pillow==7.1.2
scikit-image==0.17.2
scipy==1.5.0
scikit-learn==0.23.1
torch==1.5.1
torchvision==0.6.1
tornado~=6.0.4
tqdm==4.46.1
opencv-python==4.2.0.34
httpimport==0.7.2